requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
gunicorn==22.0.0
playwright==1.45.1
pymongo==4.8.0
python-dotenv==1.0.1
certifi==2024.7.4
dnspython==2.6.1